pyproject.toml
# ASFOTEC-MLStack-Demo pyproject.toml
[tool.poetry]
name = "asfotec-mlstack-demo"
version = "0.1.0"
description = "Showroom ML/LLM/MLOps pour ASFOTEC"
authors = ["Abselom <abselom@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.9.2"
pydantic-settings = "^2.6.0"
mlflow = "^2.20.0"
lightgbm = "^4.5.0"
scikit-learn = "^1.5.2"
langchain = "^0.3.4"
langchain-community = "^0.3.4"
langchain-google-genai = "^2.0.1"  # Gemini/Vertex
langchain-chroma = "^0.2.2"
chromadb = "^0.5.7"
pandas = "^2.2.3"
numpy = "^2.1.1"
confluent-kafka = "^2.5.0"  # Kafka producer/consumer
psycopg2-binary = "^2.9.9"  # Postgres
python-dotenv = "^1.0.1"
pytest = "^8.3.3"
ruff = "^0.7.0"
black = "^24.8.0"
boto3 = "^1.40.75"
sentence-transformers = "^5.1.2"

[tool.poetry.dev-dependencies]
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
select = ["E", "F", "B", "Q"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
